refactor(adapter-oas): drop barrel files, rename schemaHelpers, trim snapshots
- Remove emit/discriminator/index.ts and load/index.ts, importing
  directly from the concrete modules instead
- Rename emit/schemaHelpers.ts to emit/schemaShape.ts — it derives
  the AST-relevant shape of a schema (type, format, exclusive
  bounds, flattened allOf), not a grab bag of unrelated helpers
- Drop the petStore and withExternalFileRef characterization
  snapshots (9.4k and ~100 lines of generated JSON); the refactor
  they guarded is complete and reviewed, and withFragmentRef stays
  as the remaining byte-identical safety net
