+++ packages/adapter-oas/mocks/oas.ts
@@ -1,4 +1,4 @@
-import { parseDocument } from '../src/load/index.ts'
+import { parseDocument } from '../src/load/normalize.ts'
 import type { Document } from '../src/types.ts'
 
 export async function buildMinimalOas() {
