+++ packages/adapter-oas/src/adapter.ts
@@ -2,9 +2,10 @@ import { ast, findCircularSchemas, narrowSchema } from '@kubb/ast'
 import { createAdapter } from '@kubb/core'
 import type { AdapterSource } from '@kubb/core'
 import { DEFAULT_PARSER_OPTIONS } from './constants.ts'
-import { buildDiscriminatorChildMap, patchDiscriminatorNode } from './emit/discriminator/index.ts'
-import type { DiscriminatorTarget } from './emit/discriminator/index.ts'
-import { assertInputExists, parseDocument, parseFromConfig, validateDocument } from './load/index.ts'
+import { buildDiscriminatorChildMap, patchDiscriminatorNode } from './emit/discriminator/propagate.ts'
+import type { DiscriminatorTarget } from './emit/discriminator/propagate.ts'
+import { assertInputExists } from './load/source.ts'
+import { parseDocument, parseFromConfig, validateDocument } from './load/normalize.ts'
 import { getSchemas } from './model/components.ts'
 import { resolveBaseUrl } from './model/server.ts'
 import { getOperations } from './operation.ts'
