--- packages/adapter-oas/src/emit/discriminator/index.ts
@@ -1,2 +1,0 @@
-export * from './preserve.ts'
-export * from './propagate.ts'
