+++ packages/adapter-oas/src/emit/schemaShape.ts
