--- packages/adapter-oas/src/load/index.ts
@@ -1,2 +1,0 @@
-export { assertInputExists, resolveSource } from './source.ts'
-export { bundleDocument, parseDocument, parseFromConfig, validateDocument } from './normalize.ts'
