+++ packages/adapter-oas/src/parser.test.ts
@@ -3,7 +3,7 @@ import { syncSchemaRef } from '@kubb/kit'
 import { describe, expect, it } from 'vitest'
 import { buildMinimalOas } from '../mocks/oas.ts'
 import { DEFAULT_PARSER_OPTIONS } from './constants.ts'
-import { parseDocument } from './load/index.ts'
+import { parseDocument } from './load/normalize.ts'
 import { getSchemas } from './model/components.ts'
 import { getOperations } from './operation.ts'
 import { createSchemaParser, type OasParserContext } from './parser.ts'
