+++ packages/adapter-oas/src/parser.ts
@@ -2,7 +2,7 @@ import { pascalCase } from '@internals/utils'
 import { ast, type StatusCode } from '@kubb/ast'
 import { DEFAULT_PARSER_OPTIONS } from './constants.ts'
 import { type ConvertContext, schemaRules } from './emit/parseSchema.ts'
-import { flattenSchema } from './emit/schemaHelpers.ts'
+import { flattenSchema } from './emit/schemaShape.ts'
 import { getParameters, getRequestBodyContentTypes, getRequestSchema, getResponseBodyContentTypes, getResponseSchema } from './model/operations.ts'
 import { isNullable, isReference } from './oas.ts'
 import { getOperationId, getRequestBody, getRequestContentType, getResponseByStatusCode, getResponseStatusCodes } from './operation.ts'
